fixed realy date in trip-point-edit
